backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aioresponses==0.7.4
httpx==0.25.2
python-dotenv==1.0.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.3
requests==2.31.0
asyncio==3.4.3
typing-extensions==4.8.0
# External dependencies
twelvelabs==1.0.0b0
vellum-ai
requests
httpx
